+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Add validation function and associated flow [#112](https://github.com/datagouv/csv-detective/pull/112)
 - Better float detection [#113](https://github.com/datagouv/csv-detective/pull/113)
 - Refactor fields tests [#114](https://github.com/datagouv/csv-detective/pull/114)
+- Better code waldec and add code import [#116](https://github.com/datagouv/csv-detective/pull/116)
 
 ## 0.7.4 (2024-11-15)